+++ src/Enum/KindsEnum.php
@@ -48,7 +48,7 @@ enum KindsEnum: int
     case PUBLICATION_INDEX = 30040; // NKBIP-01
     case PUBLICATION_CONTENT = 30041; // NKBIP-01
     case DIRECTORY = 30045; // NKBIP-04, filesystem directory
-    case WIKI = 30817; // NIP-54 wiki entry
+    case WIKI = 30818; // NIP-54 wiki entry
     case APP_DATA = 30078; // NIP-78, Arbitrary custom app data
     case FEED_EXPRESSION = 30880; // NIP-EX, publishable feed expressions
     case PLAYLIST = 34139; // Playlist (e.g. Nostria music playlists)